Mod: `packet_conv.py` moved with test-code import fixing, for py-modules organize
